requirements.txt
-i https://pypi.org/simple
asn1crypto==1.5.1
black==26.1.0; python_version >= '3.10'
boto3==1.42.39; python_version >= '3.9'
botocore==1.42.39; python_version >= '3.9'
certifi==2026.1.4; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.3.1; python_version >= '3.10'
cryptography==46.0.4; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
filelock==3.20.3; python_version >= '3.10'
idna==3.11; python_version >= '3.8'
jaraco.classes==3.4.0; python_version >= '3.8'
jaraco.context==6.1.0; python_version >= '3.9'
jaraco.functools==4.4.0; python_version >= '3.9'
jmespath==1.1.0; python_version >= '3.9'
keyring==25.7.0; python_version >= '3.9'
more-itertools==10.8.0; python_version >= '3.9'
mypy-extensions==1.1.0; python_version >= '3.8'
numpy==2.4.1; python_version >= '3.11'
packaging==26.0; python_version >= '3.8'
pandas==3.0.0; python_version >= '3.11'
pathspec==1.0.4; python_version >= '3.9'
platformdirs==4.5.1; python_version >= '3.10'
pycparser==3.0; python_version >= '3.10'
pyjwt==2.11.0; python_version >= '3.9'
pyopenssl==25.3.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytokens==0.4.1; python_version >= '3.8'
pytz==2025.2
requests==2.32.5; python_version >= '3.9'
s3transfer==0.16.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
snowflake-connector-python[secure-local-storage]==4.2.0; python_version >= '3.9'
sortedcontainers==2.4.0
tomlkit==0.14.0; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
